# swagger_rcs_gateway.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "RCS Gateway API with Authentication",
...
        "contact": {
            "name": "Suporte RCS Gateway",
            "email": "suporte@exemplo.com"
        },
        "version": "2.0.0"
    },
    "paths": {
...
        "/api/client/send-message": {
            "post": {
                "tags": [
                    "Servi\u00e7os para Clientes"
                ],
                "summary": "Send Message",
                "description": "Endpoint principal para clientes enviarem mensagens RCS\n\nEste endpoint permite que clientes autenticados enviem mensagens RCS\nespecificando o tipo (basic ou single) e opcionalmente um template.\n\n**Par\u00e2metros:**\n- **client_code**: C\u00f3digo do cliente (deve corresponder ao cliente autenticado)\n- **user_code**: C\u00f3digo do usu\u00e1rio (opcional)\n- **message_type**: Tipo da mensagem (\"basic\" ou \"single\")\n- **template_id**: ID do template da plataforma (opcional)\n- **phone_numbers**: Lista de n\u00fameros de telefone\n- **variables**: Vari\u00e1veis para substitui\u00e7\u00e3o no template\n- **content**: Conte\u00fado personalizado (se n\u00e3o usar template)\n- **callback_url**: URL de callback (opcional, usa padr\u00e3o do cliente)\n- **campaign_name**: Nome da campanha (opcional)\n\n**Exemplo de uso:**\n```json\n{\n    \"client_code\": \"CLI_ABC123\",\n    \"user_code\": \"user001\",\n    \"message_type\": \"basic\",\n    \"template_id\": \"template_001\",\n    \"phone_numbers\": [\"5511999999999\"],\n    \"variables\": {\n        \"nome\": \"Jo\u00e3o\",\n        \"produto\": \"Smartphone\"\n    },\n    \"callback_url\": \"https://meusite.com/callback\",\n    \"campaign_name\": \"Promocao_Janeiro\"\n}\n```",
                "operationId": "send_message_api_client_send_message_post",
                "parameters": [
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "API Key do cliente",
                            "title": "X-Api-Key"
                        },
                        "description": "API Key do cliente"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClientSendMessageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClientSendMessageResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/client/messages": {
            "get": {
                "tags": [
                    "Servi\u00e7os para Clientes"
                ],
                "summary": "Get Messages",
                "description": "Consultar mensagens enviadas pelo cliente\n\nEste endpoint permite que clientes consultem suas mensagens enviadas\ncom filtros opcionais por data, status, n\u00famero de telefone, etc.\n\n**Par\u00e2metros de consulta:**\n- **client_code**: C\u00f3digo do cliente\n- **start_date**: Data inicial (formato: YYYY-MM-DD)\n- **end_date**: Data final (formato: YYYY-MM-DD)\n- **status**: Status da mensagem (pending, sent, delivered, read, failed)\n- **phone_number**: N\u00famero de telefone espec\u00edfico\n- **limit**: Limite de resultados (padr\u00e3o: 100)\n- **offset**: Offset para pagina\u00e7\u00e3o (padr\u00e3o: 0)\n\n**Exemplo de resposta:**\n```json\n[\n    {\n        \"id\": \"123\",\n        \"client_message_id\": \"CLI_ABC123_a1b2c3d4\",\n        \"phone_number\": \"5511999999999\",\n        \"status\": \"delivered\",\n        \"message_type\": \"basic\",\n        \"campaign_name\": \"Promocao_Janeiro\",\n        \"sent_at\": \"2025-01-15T10:30:00Z\",\n        \"delivered_at\": \"2025-01-15T10:31:00Z\",\n        \"created_at\": \"2025-01-15T10:30:00Z\"\n    }\n]\n```",
                "operationId": "get_messages_api_client_messages_get",
                "parameters": [
                    {
                        "name": "client_code",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Client Code"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "Start Date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "End Date"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "Status"
                        }
                    },
                    {
                        "name": "phone_number",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "Phone Number"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Offset"
                        }
                    },
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "API Key do cliente",
                            "title": "X-Api-Key"
                        },
                        "description": "API Key do cliente"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ClientMessageResponse"
                                    },
                                    "title": "Response Get Messages Api Client Messages Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/client/messages/{client_message_id}": {
            "get": {
                "tags": [
                    "Servi\u00e7os para Clientes"
                ],
                "summary": "Get Message By Id",
                "description": "Consultar mensagem espec\u00edfica pelo ID do cliente\n\n**Par\u00e2metros:**\n- **client_message_id**: ID \u00fanico da mensagem gerado para o cliente",
                "operationId": "get_message_by_id_api_client_messages__client_message_id__get",
                "parameters": [
                    {
                        "name": "client_message_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Client Message Id"
                        }
                    },
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "API Key do cliente",
                            "title": "X-Api-Key"
                        },
                        "description": "API Key do cliente"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClientMessageResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/client/stats": {
            "get": {
                "tags": [
                    "Servi\u00e7os para Clientes"
                ],
                "summary": "Get Client Stats",
                "description": "Estat\u00edsticas do cliente\n\nRetorna estat\u00edsticas b\u00e1sicas das mensagens enviadas pelo cliente.",
                "operationId": "get_client_stats_api_client_stats_get",
                "parameters": [
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "API Key do cliente",
                            "title": "X-Api-Key"
                        },
                        "description": "API Key do cliente"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ClientMessageResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "client_message_id": {
                        "type": "string",
                        "title": "Client Message Id"
                    },
                    "phone_number": {
                        "type": "string",
                        "title": "Phone Number"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "message_type": {
                        "type": "string",
                        "title": "Message Type"
                    },
                    "campaign_name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Campaign Name"
                    },
                    "sent_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Sent At"
                    },
                    "delivered_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Delivered At"
                    },
                    "read_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Read At"
                    },
                    "error_message": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Error Message"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "client_message_id",
                    "phone_number",
                    "status",
                    "message_type",
                    "created_at"
                ],
                "title": "ClientMessageResponse"
            },
            "ClientSendMessageRequest": {
                "properties": {
                    "client_code": {
                        "type": "string",
                        "title": "Client Code"
                    },
                    "user_code": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "User Code"
                    },
                    "message_type": {
                        "type": "string",
                        "title": "Message Type"
                    },
                    "template_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Template Id"
                    },
                    "phone_numbers": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Phone Numbers"
                    },
                    "variables": {
                        "anyOf": [
                            {
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Variables",
                        "default": {}
                    },
                    "content": {
                        "anyOf": [
                            {
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Content"
                    },
                    "callback_url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Callback Url"
                    },
                    "campaign_name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Campaign Name"
                    }
                },
                "type": "object",
                "required": [
                    "client_code",
                    "message_type",
                    "phone_numbers"
                ],
                "title": "ClientSendMessageRequest"
            },
            "ClientSendMessageResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "title": "Success"
                    },
                    "message_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Message Id"
                    },
                    "client_message_id": {
                        "type": "string",
                        "title": "Client Message Id"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    },
                    "sent_count": {
                        "type": "integer",
                        "title": "Sent Count"
                    },
                    "failed_count": {
                        "type": "integer",
                        "title": "Failed Count"
                    },
                    "details": {
                        "anyOf": [
                            {
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Details"
                    }
                },
                "type": "object",
                "required": [
                    "success",
                    "client_message_id",
                    "status",
                    "message",
                    "sent_count",
                    "failed_count"
                ],
                "title": "ClientSendMessageResponse"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}